eliminated circular dependency, set __main__ for beams
