+++ beams/__main__.py
@@ -1,4 +1,6 @@
 
 if __name__ == '__main__':
-    from app import beams
+    from app.beams import BEAMS
+    app = BEAMS()
+    app.run()
