+++ beams/app/resources/light_style.qss
@@ -107,7 +107,7 @@ QComboBox::drop-down {
 }
 
 QComboBox::down-arrow {
-    image: url(resources/icons/down_arrow.jpg);
+    image: url(beams/app/resources/icons/down_arrow.jpg);
 }
 
 QComboBox::down-arrow:on { /* shift the arrow when popup is open */
